+++ src/control/cmd/daos_server/network.go
@@ -79,7 +79,7 @@ func (cmd *networkScanCmd) Execute(args []string) error {
 		hf.AddInterface(&control.HostFabricInterface{
 			Provider: fi.Provider,
 			Device:   fi.DeviceName,
-			NumaNode: fi.NUMANode,
+			NumaNode: uint32(fi.NUMANode),
 		})
 	}
